move out App component from index.js to App.js
